refactor: Rename localStorage to useLocalStorage
