+++ frontend/src/hooks/index.ts
@@ -1,2 +1,2 @@
-export * from './localStorage';
+export * from './useLocalStorage';
 export * from './user';
